+++ web/src/modules/Student/Single/Fees/StudentLedgerPage.tsx
@@ -70,10 +70,18 @@ export const StudentLedgerPage: React.FC<StudentLedgerPageProp> = ({
 		// get class additional fees
 		const classAdditionals = settings?.classes?.additionalFees?.[section.class_id]
 
-		fees = {
-			...fees,
-			[student.id]: classFee,
-			...classAdditionals
+		if (classFee) {
+			fees = {
+				...fees,
+				[student.id]: classFee
+			}
+		}
+
+		if (classAdditionals) {
+			fees = {
+				...fees,
+				...classAdditionals
+			}
 		}
 	}
 
@@ -336,7 +344,10 @@ export const getMergedFees = (
 
 				return {
 					...agg,
-					[id]: {
+					// this is to make sure, all fees should have unique ID
+					// because 2 or more than 2 students can have same fee or payment ID
+					// if they're are in the same class (now we're generating payments from class additionals)
+					[id + '$' + student.id]: {
 						...fee
 					}
 				}