+++ web/src/pages/class/create.tsx
@@ -58,7 +58,7 @@ export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> 
 	const handleInput = (
 		event: React.ChangeEvent<HTMLInputElement & HTMLSelectElement & HTMLTextAreaElement>
 	) => {
-		const { name, value, type } = event.target
+		const { name, value, type, valueAsNumber } = event.target
 
 		if (name === 'name') {
 			return setState({
@@ -76,7 +76,7 @@ export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> 
 				...state,
 				class: {
 					...state.class,
-					[name]: parseInt(value)
+					[name]: parseInt(value) > 0 ? valueAsNumber : 0
 				}
 			})
 		}
@@ -153,11 +153,8 @@ export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> 
 	}
 
 	return (
-		<AppLayout title={isNewClass ? 'Add New Class' : 'Update Class'} showHeaderTitle>
+		<AppLayout title={isNewClass ? 'Add new Class' : 'Update Class'} showHeaderTitle>
 			<div className="p-5 md:p-10 md:pt-5 md:pb-0 relative">
-				{/* <div className="text-2xl font-bold mt-4 mb-8 text-center">
-					{isNewClass ? 'Add New Class' : 'Update Class'}
-				</div> */}
 				<div className="md:w-4/5 md:mx-auto flex flex-col items-center space-y-3 rounded-2xl bg-gray-700 py-5 my-4">
 					<div className="text-white text-center text-base my-5 font-semibold">
 						Fill Class Information