+++ web/src/pages/school/fees/print-voucher/preview.tsx
@@ -7,7 +7,7 @@ import chunkify from 'utils/chunkify'
 import getFilteredPayments from 'utils/getFilteredPayments'
 import getSectionFromId from 'utils/getSectionFromId'
 import { StudentLedgerPage } from 'modules/Student/Single/Fees/StudentLedgerPage'
-import { isValidStudent } from 'utils'
+import { isValidStudent, rollNumberSorter } from 'utils'
 
 import './style.css'
 
@@ -32,7 +32,7 @@ const PrintPreview = () => {
 
 	const siblings = (): AugmentedSibling[] => {
 		return Object.values(students)
-			.filter(s => s && s.Name && s.FamilyID && s.FamilyID === id)
+			.filter(s => s && s.Active && s.Name && s.FamilyID && s.FamilyID === id)
 			.reduce((agg, curr) => {
 				const section_id = curr.section_id
 				return [
@@ -58,8 +58,12 @@ const PrintPreview = () => {
 		// separate logic to print
 		if (type === 'CLASS') {
 			return Object.values(students)
-				.filter(s => isValidStudent(s) && currClass.sections[s.section_id] !== undefined)
-				.sort((a, b) => parseInt(a.RollNumber || '0') - parseInt(b.RollNumber || '0'))
+				.filter(
+					s =>
+						isValidStudent(s, { active: true }) &&
+						currClass.sections[s.section_id] !== undefined
+				)
+				.sort(rollNumberSorter)
 		}
 
 		return [getCurrStudent()]
@@ -67,8 +71,8 @@ const PrintPreview = () => {
 
 	const relevantStudents = getRelevantStudents()
 
-	const getMergedPaymentsForStudents = (student: MISStudent) => {
-		if (id !== undefined && type === 'FAMILY') {
+	const getMergedPaymentsForStudents = (student?: MISStudent) => {
+		if (id && type === 'FAMILY') {
 			const familyStudents = siblings()
 			const merged_payments = familyStudents.reduce(
 				(agg, curr) => ({
@@ -76,9 +80,12 @@ const PrintPreview = () => {
 					...Object.entries(curr.payments ?? {}).reduce((agg, [pid, p]) => {
 						return {
 							...agg,
-							[pid]: {
+							// // this is to make sure, all payments should have unique ID
+							// because 2 or more than 2 siblings can have same fee or payment ID
+							// if they're are in the same class (now we're generating payments from class additionals)
+							[pid + '$' + curr.id]: {
 								...p,
-								fee_name: p.fee_name && `${curr.Name}-${p.fee_name}`
+								fee_name: `${curr.Name}-${p.fee_name}`
 							}
 						}
 					}, {} as MISStudent['payments'])
@@ -119,7 +126,7 @@ const PrintPreview = () => {
 							<StudentLedgerPage
 								key={student.id}
 								payments={getFilteredPayments(
-									getMergedPaymentsForStudents(student),
+									getMergedPaymentsForStudents(),
 									'',
 									''
 								)}
@@ -171,11 +178,7 @@ const PrintPreview = () => {
 					vouchers.push(
 						<StudentLedgerPage
 							key={student.id + i} // adding i to avoid key duplicaiton
-							payments={getFilteredPayments(
-								getMergedPaymentsForStudents(student),
-								'',
-								''
-							)}
+							payments={getFilteredPayments(getMergedPaymentsForStudents(), '', '')}
 							settings={settings}
 							family={getFamily()}
 							voucherNo={voucher_no}