+++ web/src/pages/students/add/index.tsx
@@ -13,7 +13,7 @@ import toTitleCase from 'utils/toTitleCase'
 import getSectionsFromClasses from 'utils/getSectionsFromClasses'
 import AdmissionForm from 'components/Printable/Student/admissionform'
 import { isValidPhone } from 'utils/helpers'
-import { formatCNIC } from 'utils'
+import { classYearSorter, formatCNIC } from 'utils'
 import { getImageString, getDownsizedImage } from 'utils/image'
 
 import { UploadImage } from 'components/image'
@@ -363,11 +363,10 @@ export const CreateOrUpdateStudent = () => {
 							className="w-full tw-select">
 							<option value={''}>Choose</option>
 							{sections
-								.sort((a, b) => (a.classYear ?? 0) - (b.classYear ?? 0))
-								.filter(s => s && s.id && s.name)
+								.sort(classYearSorter)
 								.map(s => (
 									<option key={s.id} value={s.id}>
-										{toTitleCase(s.namespaced_name, '-')}
+										{s.namespaced_name}
 									</option>
 								))}
 						</select>